+++ spec/head_music/style/guidelines/notes_same_length_spec.rb
@@ -58,15 +58,12 @@ describe HeadMusic::Style::Guidelines::NotesSameLength do
       voice.place("2:1", :half, "E4")
       voice.place("2:3", :half, "F4")
       voice.place("3:1", :whole, "E4")
-      voice.place("2:1", :half, "E4")
-      voice.place("2:3", :half, "F4")
-      voice.place("3:1", :whole, "E4")
       voice.place("4:1", :whole, "D4")
     end
 
     its(:fitness) { is_expected.to be < 1 }
     its(:message) { is_expected.not_to be_empty }
-    its(:marks_count) { is_expected.to eq 4 }
+    its(:marks_count) { is_expected.to eq 2 }
     its(:first_mark_code) { is_expected.to eq "2:1:000 to 2:3:000" }
   end
 end